refactor(core): Move node type policy DTOs to @n8n/api-types
Move the policy rule schemas and the internal request DTOs from
packages/cli to @n8n/api-types. The public API surface for node type
policies needs its DTOs there, per the public API convention that
request/response DTOs live in @n8n/api-types.

Behavior-neutral: only import paths change. Split out of the node type
policy public API endpoints work (IAM-1144) so that work lands in
smaller, easier-to-review pieces.
