+++ packages/@n8n/api-types/src/dto/index.ts
@@ -220,6 +220,27 @@ export {
 } from './log-streaming/public-destination.dto';
 
 export { PaginationDto, MAX_ITEMS_PER_PAGE } from './pagination/pagination.dto';
+
+export type {
+	NonDelegatingPolicyAction,
+	NonDelegatingPolicyRule,
+	PolicyAction,
+	PolicyRule,
+	PolicySelector,
+} from './node-type-policies/policy-rule.types';
+export { policySelectorSchema } from './node-type-policies/policy-selector.schema';
+export {
+	nonDelegatingPolicyActionSchema,
+	nonDelegatingPolicyRuleListSchema,
+	policyActionSchema,
+	policyRuleListSchema,
+	policyRuleSchema,
+} from './node-type-policies/policy-rule.schema';
+export { PutInstancePolicyDto } from './node-type-policies/put-instance-policy.dto';
+export { PutProjectPolicyDto } from './node-type-policies/put-project-policy.dto';
+export { CreatePolicyDocumentDto } from './node-type-policies/create-policy-document.dto';
+export { UpdatePolicyDocumentDto } from './node-type-policies/update-policy-document.dto';
+export { ReplaceAttachmentsDto } from './node-type-policies/replace-attachments.dto';
 export {
 	UsersListFilterDto,
 	type UsersListSortOptions,
