+++ packages/@n8n/api-types/src/dto/node-type-policies/create-policy-document.dto.ts
@@ -1,6 +1,5 @@
-import { Z } from '@n8n/api-types';
-
 import { policyRuleListSchema } from './policy-rule.schema';
+import { Z } from '../../zod-class';
 
 /** Request body for creating a policy document. */
 export class CreatePolicyDocumentDto extends Z.class({
