+++ packages/@n8n/api-types/src/dto/node-type-policies/policy-selector.schema.ts
@@ -1,9 +1,9 @@
 import { z } from 'zod';
 
-import type { PolicySelector } from '../policy-rule.types';
+import type { PolicySelector } from './policy-rule.types';
 
 /**
- * Zod counterpart of `PolicySelector` in `../policy-rule.types.ts`. The `satisfies` check
+ * Zod counterpart of `PolicySelector` in `./policy-rule.types.ts`. The `satisfies` check
  * keeps this schema honest against that type — it fails to compile if this schema ever
  * accepts a shape `PolicySelector` doesn't.
  */
